web/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bs4==0.0.2
Flask==3.0.3
numpy-financial==1.0.0
openpyxl==3.1.2
python-dotenv==1.0.1
requests==2.31.0
Unidecode==1.3.8
